coverage only for the first evaluated version

+++ Jenkinsfile
@@ -43,6 +43,7 @@ def getWheelPath(tox_skip_install, python_version) {
 }
 
 def runTest(protocol, slave = 0, tox_skip_install = false) {
+    def firstIteration = true
     def pythonVersions = RUN_PYTHON_VERSIONS.split(',')
     pythonVersions.each { version ->
         def wheelFile = getWheelPath(tox_skip_install, version)
@@ -58,13 +59,16 @@ def runTest(protocol, slave = 0, tox_skip_install = false) {
         } catch (err) {
             unstable(message: "Tests failed")
         } finally {
-            def coverage_stash = ".coverage_${protocol}_${slave}"
-            bat "move .coverage ${coverage_stash}"
-            junit "pytest_reports\\*.xml"
-            // Delete the junit after publishing it so it not re-published on the next stage
-            bat "del /S /Q pytest_reports\\*.xml"
-            stash includes: coverage_stash, name: coverage_stash
-            coverage_stashes.add(coverage_stash)
+            if (firstIteration) {
+                def coverage_stash = ".coverage_${protocol}_${slave}"
+                bat "move .coverage ${coverage_stash}"
+                junit "pytest_reports\\*.xml"
+                // Delete the junit after publishing it so it not re-published on the next stage
+                bat "del /S /Q pytest_reports\\*.xml"
+                stash includes: coverage_stash, name: coverage_stash
+                coverage_stashes.add(coverage_stash)
+                firstIteration = false
+            }
         }
     }
 }
